moved function back to main
